refactor(components): 共有 component の story を packages/components に colocate
Storybook の Components を packages/components/src の中身と揃えるため、
packages の component を題材にした story（28 本）と AlertDialog の docs mdx を
product から packages 側へ移設し、component とコロケーションする。

- 単一 barrel の self-import（@dayopt/components）が package 内 tsc で解決する
  ことを確認の上、各 category の story を packages/components/src/{category}/ へ移動
- packages/components に storybook toolchain を devDep 追加
  （@storybook/nextjs-vite / @storybook/addon-docs / storybook）
- app 依存を持つ 3 story を移設可能化:
  - checkbox: @/features/tags の getTagColorClasses を foundations トークン
    リテラル（var(--tag-green)）に置換
  - separator: @dayopt/storybook の withWrapper decorator を inline 化
  - avatar: 基底 Avatar story を packages へ、app 固有の AvatarUpload demo は
    product の avatar-upload.stories として分離
- alert-dialog.docs.mdx を packages へ移し、storybook main.ts の stories glob に
  packages/components の mdx を追加
- app 固有 component の story（toast/mini-calendar/confirm-dialog/EmptyState 等）と
  props-inventory は product に据え置き

typecheck / lint / lint:boundaries / 全 1848 test / storybook build pass。
