+++ packages/components/package.json
@@ -44,8 +44,11 @@
     "vaul": "^1.1.2"
   },
   "devDependencies": {
+    "@storybook/addon-docs": "^10.4.6",
+    "@storybook/nextjs-vite": "^10.4.6",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "storybook": "^10.4.6",
     "typescript": "^5"
   }
 }
