+++ packages/components/src/display/avatar.stories.tsx
@@ -1,9 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/nextjs-vite';
-import { useState } from 'react';
 
 import { Avatar, AvatarFallback, AvatarImage } from '@dayopt/components';
-
-import { AvatarUpload } from '../inputs/avatar-upload';
 
 const meta = {
   title: 'Components/UI/Display/Avatar',
@@ -14,10 +11,6 @@ const meta = {
 
 export default meta;
 type Story = StoryObj<typeof meta>;
-
-// ---------------------------------------------------------------------------
-// Stories
-// ---------------------------------------------------------------------------
 
 /** サイドバー用（xs相当 24px） */
 export const Default: Story = {
@@ -37,10 +30,6 @@ export const AccountSettings: Story = {
     </Avatar>
   ),
 };
-
-// ---------------------------------------------------------------------------
-// AllPatterns
-// ---------------------------------------------------------------------------
 
 /** 全パターン一覧 */
 export const AllPatterns: Story = {
@@ -87,52 +76,4 @@ export const AllPatterns: Story = {
       </section>
     </div>
   ),
-};
-
-// ---------------------------------------------------------------------------
-// AvatarUpload
-// ---------------------------------------------------------------------------
-
-function AvatarUploadDemo({ initialUrl }: { initialUrl?: string }) {
-  const [isUploading, setIsUploading] = useState(false);
-  const [avatarUrl, setAvatarUrl] = useState<string | null>(initialUrl || null);
-
-  const handleUpload = async (file: File) => {
-    setIsUploading(true);
-    await new Promise((resolve) => setTimeout(resolve, 1000));
-    const url = URL.createObjectURL(file);
-    setAvatarUrl(url);
-    setIsUploading(false);
-  };
-
-  const handleRemove = async () => {
-    setIsUploading(true);
-    await new Promise((resolve) => setTimeout(resolve, 500));
-    setAvatarUrl(null);
-    setIsUploading(false);
-  };
-
-  return (
-    <AvatarUpload
-      currentAvatarUrl={avatarUrl}
-      onUpload={handleUpload}
-      onRemove={handleRemove}
-      loading={isUploading}
-      size="3xl"
-    />
-  );
-}
-
-/** アップロード（空状態） */
-export const Upload: Story = {
-  // label: hidden file input inside AvatarUpload component
-  parameters: { a11y: { test: 'todo' } },
-  render: () => <AvatarUploadDemo />,
-};
-
-/** アップロード（画像あり） */
-export const UploadWithImage: Story = {
-  // label: hidden file input inside AvatarUpload component
-  parameters: { a11y: { test: 'todo' } },
-  render: () => <AvatarUploadDemo initialUrl="https://github.com/shadcn.png" />,
 };
